refactor: pull ios logic into own file, build multiple architectures
