+++ test-e2e/src/bin/ios.rs
@@ -1,3 +1,3 @@
 fn main() {
-    mopro_ffi::app_config::build();
+    mopro_ffi::app_config::ios::build();
 }
